+++ apps/web/src/app/(main)/(docs)/components/autocomplete/variants/disabled.tsx
@@ -59,20 +59,18 @@ const animals = [
   },
 ];
 
-export const AutocompleteUsage = (props: Partial<AutocompleteProps>) => {
+export const AutocompleteDisabled = (props: Partial<AutocompleteProps>) => {
   return (
-    <div>
-      <Autocomplete label="Animal" {...props}>
-        {animals.map((animal) => (
-          <AutocompleteItem
-            key={animal.value}
-            value={animal.value}
-            label={animal.label}
-          >
-            {animal.label}
-          </AutocompleteItem>
-        ))}
-      </Autocomplete>
-    </div>
+    <Autocomplete isDisabled label="Animal" {...props}>
+      {animals.map((animal) => (
+        <AutocompleteItem
+          key={animal.value}
+          value={animal.value}
+          label={animal.label}
+        >
+          {animal.label}
+        </AutocompleteItem>
+      ))}
+    </Autocomplete>
   );
 };