+++ packages/components/autocomplete/src/use-autocomplete-context.ts
@@ -1,12 +1,14 @@
 import { type useInteractions } from "@floating-ui/react";
 import { createContext } from "@jamsr-ui/utils";
 import type { SelectionSet } from "./use-autocomplete";
+import { type autocompleteVariant } from "./style";
 
 export type AutocompleteContextType = {
   activeIndex: number | null;
   getItemProps: ReturnType<typeof useInteractions>["getItemProps"];
   handleSelect: (args: { value: string; label: string }) => void;
   value: SelectionSet;
+  styles: ReturnType<typeof autocompleteVariant>;
 };
 
 export const [AutocompleteProvider, useAutocompleteContext] =