+++ packages/components/rhf/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @jamsr-ui/rhf
 
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/autocomplete@0.0.95
+  - @jamsr-ui/input@0.0.89
+  - @jamsr-ui/editor@0.0.93
+  - @jamsr-ui/tags-input@0.0.25
+  - @jamsr-ui/textarea@0.0.89
+
 ## 0.0.32
 
 ### Patch Changes